refactor(react): create button dir
