+++ packages/react/src/components/button/index.tsx
@@ -1,0 +1,2 @@
+export * from './Button'
+export * from './parts'
