+++ packages/react/src/components/button/parts.ts
@@ -1,0 +1,19 @@
+import { Button } from './Button'
+
+/**
+ * This module contains the parts of the Button component.
+ * @module 'button/parts'
+ */
+
+/**
+ * An Object containing the parts of the Button component. For users that
+ * prefer Object component syntax.
+ *
+ * @remarks
+ *
+ * When using object component syntax, you import the ButtonParts object and
+ * the entire family of components vs. only what you use.
+ */
+export const ButtonParts = {
+  Root: Button,
+}
