+++ packages/react/src/index.ts
@@ -20,7 +20,7 @@ export * from './components/Accordion.client'
 export * from './components/AccordionItemGroup'
 export * from './components/Admonition'
 export * from './components/Avatar'
-export * from './components/Button'
+export * from './components/button/index'
 export * from './components/Checkbox'
 export * from './components/CircularProgress'
 export * from './components/DatePicker.client'
